Move tests to handlers_test.go
